Moved listDisksInWorkspace into an admin controller.
