+++ api/src/main/java/org/pmiops/workbench/api/DisksController.java
@@ -3,12 +3,9 @@ package org.pmiops.workbench.api;
 import java.util.List;
 import java.util.logging.Logger;
 import java.util.stream.Collectors;
-import org.pmiops.workbench.annotations.AuthorityRequired;
-import org.pmiops.workbench.disks.DiskService;
 import org.pmiops.workbench.leonardo.LeonardoApiClient;
 import org.pmiops.workbench.leonardo.PersistentDiskUtils;
 import org.pmiops.workbench.leonardo.model.LeonardoListPersistentDiskResponse;
-import org.pmiops.workbench.model.Authority;
 import org.pmiops.workbench.model.Disk;
 import org.pmiops.workbench.model.DiskStatus;
 import org.pmiops.workbench.model.EmptyResponse;
@@ -26,18 +23,15 @@ public class DisksController implements DisksApiDelegate {
   private final LeonardoApiClient leonardoNotebooksClient;
   private final LeonardoMapper leonardoMapper;
   private final WorkspaceService workspaceService;
-  private final DiskService diskService;
 
   @Autowired
   public DisksController(
       LeonardoApiClient leonardoNotebooksClient,
       LeonardoMapper leonardoMapper,
-      WorkspaceService workspaceService,
-      DiskService diskService) {
+      WorkspaceService workspaceService) {
     this.leonardoNotebooksClient = leonardoNotebooksClient;
     this.leonardoMapper = leonardoMapper;
     this.workspaceService = workspaceService;
-    this.diskService = diskService;
   }
 
   @Override
@@ -89,15 +83,6 @@ public class DisksController implements DisksApiDelegate {
     ListDisksResponse listDisksResponse = new ListDisksResponse();
     listDisksResponse.addAll(diskList);
 
-    return ResponseEntity.ok(listDisksResponse);
-  }
-
-  @Override
-  @AuthorityRequired({Authority.RESEARCHER_DATA_VIEW})
-  public ResponseEntity<ListDisksResponse> listDisksInWorkspace(String workspaceNamespace) {
-    List<Disk> diskList = diskService.findByWorkspaceNamespace(workspaceNamespace);
-    ListDisksResponse listDisksResponse = new ListDisksResponse();
-    listDisksResponse.addAll(diskList);
     return ResponseEntity.ok(listDisksResponse);
   }
 }
