+++ api/src/main/resources/workbench-api.yaml
@@ -4257,7 +4257,7 @@ paths:
     - "$ref": "#/parameters/workspaceNamespace"
     get:
       tags:
-      - disks
+      - diskAdmin
       description:  Lists persistent disks in a workspace. Requires the
         RESEARCHER_DATA_VIEW authority.
       operationId: listDisksInWorkspace
